drawer: extract tasks tab
